- 수행 목록 -- list형태로 받는건 전부 /{pageIndex} 로 받아서 처리하는 형태로 변경함.

+++ src/main/java/com/jhsfully/reservation/domain/Review.java
@@ -30,7 +30,7 @@ public class Review {
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
-    public static ReviewDto.ReviewResponse toDto(Review review, int reviewCount){
+    public static ReviewDto.ReviewResponse toDto(Review review, long reviewCount){
         return ReviewDto.ReviewResponse.builder()
                 .id(review.getId())
                 .reviewCount(reviewCount)
